Fixing overflow bug

+++ impl/referenceProblem.c
@@ -53,7 +53,7 @@ static ReferenceIntervalInsertion *referenceIntervalInsertion_construct(stIntTup
 }
 
 static double referenceIntervalInsertion_isConnectedLeft(ReferenceIntervalInsertion *referenceIntervalInsertion, float *z,
-        int32_t nodeNumber) {
+        int64_t nodeNumber) {
     int32_t _5Node = stIntTuple_getPosition(referenceIntervalInsertion->chain, referenceIntervalInsertion->orientation ? 0 : 1);
     assert(referenceIntervalInsertion->referenceInterval != NULL);
     assert(z[referenceIntervalInsertion->referenceInterval->_3Node * nodeNumber + _5Node] >= 0.0);
@@ -73,7 +73,7 @@ static int32_t referenceInterval_get5Node(ReferenceInterval *referenceInterval) 
 }
 
 static double referenceIntervalInsertion_isConnectedRight(ReferenceIntervalInsertion *referenceIntervalInsertion, float *z,
-        int32_t nodeNumber) {
+        int64_t nodeNumber) {
     int32_t _3Node = stIntTuple_getPosition(referenceIntervalInsertion->chain, referenceIntervalInsertion->orientation ? 1 : 0);
     int32_t _5Node = referenceInterval_get5Node(referenceIntervalInsertion->referenceInterval);
     assert(z[_5Node * nodeNumber + _3Node] >= 0.0);
@@ -81,7 +81,7 @@ static double referenceIntervalInsertion_isConnectedRight(ReferenceIntervalInser
 }
 
 static double referenceIntervalInsertion_isCurrentlyConnected(ReferenceIntervalInsertion *referenceIntervalInsertion, float *z,
-        int32_t nodeNumber) {
+        int64_t nodeNumber) {
     assert(referenceIntervalInsertion->referenceInterval != NULL);
     int32_t _3Node = referenceIntervalInsertion->referenceInterval->_3Node;
     int32_t _5Node = referenceInterval_get5Node(referenceIntervalInsertion->referenceInterval);
@@ -90,13 +90,13 @@ static double referenceIntervalInsertion_isCurrentlyConnected(ReferenceIntervalI
 }
 
 static double referenceIntervalInsertion_connectionScore(ReferenceIntervalInsertion *referenceIntervalInsertion, float *z,
-        int32_t nodeNumber) {
+        int64_t nodeNumber) {
     return referenceIntervalInsertion_isConnectedLeft(referenceIntervalInsertion, z, nodeNumber)
             + referenceIntervalInsertion_isConnectedRight(referenceIntervalInsertion, z, nodeNumber)
             - referenceIntervalInsertion_isCurrentlyConnected(referenceIntervalInsertion, z, nodeNumber);
 }
 
-static ReferenceIntervalInsertion getMaxReferenceIntervalInsertion(stList *reference, stIntTuple *chain, float *z, int32_t nodeNumber,
+static ReferenceIntervalInsertion getMaxReferenceIntervalInsertion(stList *reference, stIntTuple *chain, float *z, int64_t nodeNumber,
         double *positiveScores, double *negativeScores) {
     /*
      * Calculates the scores of the inserting the chain at each possible position in the reference.
@@ -174,7 +174,7 @@ static ReferenceInterval *insert(ReferenceIntervalInsertion *referenceIntervalIn
     return newInterval;
 }
 
-stList *makeReferenceGreedily(stList *stubs, stList *chainsList, float *z, int32_t nodeNumber, double *totalScore, bool fast) {
+stList *makeReferenceGreedily(stList *stubs, stList *chainsList, float *z, int64_t nodeNumber, double *totalScore, bool fast) {
     /*
      * Constructs a reference greedily, by picking a best possible insertion
      * at each of the |R| steps.
@@ -291,7 +291,7 @@ void greedyImprovement(stList *reference, stList *chains, float *z, int32_t perm
     /*
      * Update a reference by sampling.
      */
-    int32_t nodeNumber = (stList_length(reference) + stList_length(chains)) * 2;
+    int64_t nodeNumber = (stList_length(reference) + stList_length(chains)) * 2;
     chains = stList_copy(chains, NULL);
     double *positiveScores = st_malloc(nodeNumber * sizeof(double));
     double *negativeScores = st_malloc(nodeNumber * sizeof(double));
@@ -329,7 +329,7 @@ stList *convertReferenceToAdjacencyEdges(stList *reference) {
     return edges;
 }
 
-double calculateMaxZ(int32_t nodeNumber, float *z) {
+double calculateMaxZ(int64_t nodeNumber, float *z) {
     double maxZ = 0.0;
     for (int32_t i = 0; i < nodeNumber; i++) {
         for (int32_t j = i + 1; j < nodeNumber; j++) {
@@ -341,7 +341,7 @@ double calculateMaxZ(int32_t nodeNumber, float *z) {
     return maxZ;
 }
 
-double calculateZScoreOfReference(stList *reference, int32_t nodeNumber, float *zMatrix) {
+double calculateZScoreOfReference(stList *reference, int64_t nodeNumber, float *zMatrix) {
     double totalScore = 0.0;
     for (int32_t i = 0; i < stList_length(reference); i++) {
         ReferenceInterval *referenceInterval = stList_get(reference, i);
@@ -359,7 +359,7 @@ double calculateZScoreOfReference(stList *reference, int32_t nodeNumber, float *
     return totalScore;
 }
 
-void logZScoreOfReference(stList *reference, int32_t nodeNumber, float *zMatrix) {
+void logZScoreOfReference(stList *reference, int64_t nodeNumber, float *zMatrix) {
     for (int32_t i = 0; i < stList_length(reference); i++) {
         ReferenceInterval *referenceInterval = stList_get(reference, i);
         int32_t j = referenceInterval->_5Node;
@@ -377,7 +377,7 @@ void logZScoreOfReference(stList *reference, int32_t nodeNumber, float *zMatrix)
     }
 }
 
-void logReference(stList *reference, int32_t nodeNumber, float *zMatrix, double totalScore, const char *message) {
+void logReference(stList *reference, int64_t nodeNumber, float *zMatrix, double totalScore, const char *message) {
     st_logDebug("Reporting reference with %i intervals, %i nodes and %lf total score out of max score of %lf for %s\n",
             stList_length(reference), nodeNumber, totalScore, calculateMaxZ(nodeNumber, zMatrix), message);
     for (int32_t i = 0; i < stList_length(reference); i++) {
@@ -418,7 +418,7 @@ double constantTemperatureFn(double d) {
  * The following code is redundant to some of that above..
  */
 
-static stList *getReferenceIntervalInsertions(stList *reference, stIntTuple *chain, float *z, int32_t nodeNumber) {
+static stList *getReferenceIntervalInsertions(stList *reference, stIntTuple *chain, float *z, int64_t nodeNumber) {
     /*
      * Calculates the scores of the inserting the chain at each possible position in the reference.
      *
@@ -484,7 +484,7 @@ void gibbsSamplingWithSimulatedAnnealing(stList *reference, stList *chains, floa
     /*
      * Update a reference by sampling.
      */
-    int32_t nodeNumber = (stList_length(reference) + stList_length(chains)) * 2;
+    int64_t nodeNumber = (stList_length(reference) + stList_length(chains)) * 2;
     chains = stList_copy(chains, NULL);
     stHash *referenceIntervalToChainHash = getChainToReferenceIntervalHash(reference);
     for (int32_t k = 0; k < permutations; k++) {
